Extracted GlfwBuilder from Builder
