+++ src/examples/boxes/main.go
@@ -3,7 +3,7 @@ package main
 import . "display"
 
 func CreateBoxesApp(title string) (Displayable, error) {
-	return NewBuilder(WindowTitle(title), WindowSize(640, 480)).Build(func(b Builder) {
+	return NewGlfwBuilder(WindowTitle(title), WindowSize(640, 480)).Build(func(b Builder) {
 		Sprite(b, Children(func() {
 			Sprite(b, FlexWidth(1), FlexHeight(1), MaxWidth(640), MaxHeight(480))
 			Sprite(b, FlexWidth(1), FlexHeight(1), MaxWidth(320), MaxHeight(240))
